% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{createFormApp}
\alias{createFormApp}
\title{Creates a shinyform app from the defined questions and parameters set in formInfo.}
\usage{
createFormApp(formInfo)
}
\arguments{
\item{formInfo}{A list with param: id, questions and storage}
}
\description{
Creates a shinyform app from the defined questions and parameters set in formInfo.
}
\examples{
if (interactive()) {

library(shiny)
library(shinyforms)

questions <- list(
  list(id = "name", type = "text", title = "Name", mandatory = TRUE),
  list(id = "age", type = "numeric", title = "Age"),
  list(id = "favourite_pkg", type = "text", title = "Favourite R package"),
  list(id = "terms", type = "checkbox", title = "I agree to the terms")
)
formInfo <- list(
  id = "basicinfo",
  questions = questions,
  storage = list(
    # Right now, only flat file storage is supported
    type = STORAGE_TYPES$FLATFILE,
    # The path where responses are stored
    path = "responses"
  )
)

createFormApp(formInfo)
}
}
