% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{createFormInfo}
\alias{createFormInfo}
\title{createFormInfo}
\usage{
createFormInfo(id, questions, storage, name, multiple = TRUE, password)
}
\arguments{
\item{id}{}

\item{questions}{}

\item{storage}{}

\item{name}{}

\item{multiple}{}
}
\description{

}
\keyword{internal}
