% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{inlineInput}
\alias{inlineInput}
\title{inlineInput}
\usage{
inlineInput(tag)
}
\arguments{
\item{tag}{defined inline object}
}
\description{
Allows inline inputs to be entered i.e. for passwords and username of admins
}
\keyword{internal}
