% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{labelMandatory}
\alias{labelMandatory}
\title{labelMandatory}
\usage{
labelMandatory(label)
}
\arguments{
\item{label}{A string representing the mandatory question.}
}
\value{
A tag list
}
\description{
Adds a mandaotory star to a labelled question.
}
\keyword{internal}
