% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{loadData}
\alias{loadData}
\title{loadData}
\usage{
loadData(storage)
}
\arguments{
\item{storage}{A list with variable type defining users perferred type of storage}
}
\description{
Passes data from a storage type and passes it back to shiny.
}
\keyword{internal}
