% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{loadDataFlatfile}
\alias{loadDataFlatfile}
\title{loadDataFlatfile}
\usage{
loadDataFlatfile(storage)
}
\arguments{
\item{storage}{A list with variable type defining users perferred type of storage}
}
\description{
Takes data from a flat file and passes it to your shiny app.
}
\keyword{internal}
