% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{loadDataGsheets}
\alias{loadDataGsheets}
\title{loadDataGsheets}
\usage{
loadDataGsheets()
}
\arguments{
\item{storage}{A list with variable type defining users perferred type of storage and storage key}
}
\description{
Takes data from a google doc file and passes it to your shiny app.
}
\keyword{internal}
