% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{saveData}
\alias{saveData}
\title{saveData}
\usage{
saveData(data, storage)
}
\arguments{
\item{data}{Dataframe taken from input shiny object.}

\item{storage}{A list with variable type defining users perferred type of storage}
}
\description{
Takes data from your shinyforms inputs and passes it to a storage type.
}
\details{
Currently only provides storage of flat files (.csv) or google doc sheet.
}
\keyword{internal}
