% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{saveDataGsheets}
\alias{saveDataGsheets}
\title{saveDataGsheets}
\usage{
saveDataGsheets(data, storage)
}
\arguments{
\item{data}{Dataframe taken from input shiny object}

\item{storage}{A list with variable type defining users perferred type of storage}
}
\description{
Takes data from your shinyforms inputs and saves it to a google doc file
}
\keyword{internal}
