% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\docType{data}
\name{STORAGE_TYPES}
\alias{STORAGE_TYPES}
\title{STORAGE_TYPES}
\format{An object of class \code{list} of length 7.}
\usage{
STORAGE_TYPES
}
\value{
A list of storage types available to shinyform.
}
\description{
A list of all the available storage types for shinyforms.
}
\details{
List: FLATFILE, SQLITE, MYSQL, MONGO, GOOGLE_SHEETS, DROPBOX, AMAZON_S3
}
\keyword{datasets}
