% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{createFormInfo}
\alias{createFormInfo}
\title{createFormInfo}
\usage{
createFormInfo(id, questions, storage, name, multiple = TRUE, password)
}
\arguments{
\item{id}{String name of the form}

\item{questions}{list of form questions}

\item{storage}{a list of different storage types, path and keys}

\item{name}{String name of the app}

\item{multiple}{boolean}
}
\description{
Created a yaml file for configuring shinyforms
}
\keyword{internal}
