% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{labelMandatory}
\alias{labelMandatory}
\title{Adds a tag list with the inputted label as a parameter.}
\usage{
labelMandatory(label)
}
\arguments{
\item{label}{A string.}
}
\value{
A tag list
}
\description{
Adds a tag list with the inputted label as a parameter.
}
\keyword{internal}
